[Refactor] Gif 생성 버튼 코드 분리
