+++ src/snowEffect.js
@@ -1,5 +1,5 @@
 import React, { useRef, useEffect, useState } from "react";
-import GIF from "gif.js";
+import GenerateGIFButton from "./components/button/GenerateGIFButton";
 
 const SnowEffectOnUploadedImage = () => {
   const canvasRef = useRef(null);
@@ -124,56 +124,6 @@ const SnowEffectOnUploadedImage = () => {
     };
   }
 
-  const generateGIF = () => {
-    if (!canvasRef.current || !image) return;
-
-    const gif = new GIF({
-      workers: 2,
-      quality: 10,
-      width: imageSize.width,
-      height: imageSize.height,
-    });
-
-    const ctx = canvasRef.current.getContext("2d");
-    const snowflakes = createSnowflakes();
-
-    gif.on("finished", function (blob) {
-      const url = URL.createObjectURL(blob);
-
-      // 다운로드 링크 생성
-      const downloadLink = document.createElement("a");
-      downloadLink.href = url;
-      downloadLink.download = "snow_effect.gif"; // 다운로드되는 파일의 이름
-      downloadLink.text = "Download GIF"; // 링크 텍스트
-      downloadLink.style.display = "block"; // 링크를 보이게 설정
-      document.body.appendChild(downloadLink); // 링크를 문서에 추가
-    });
-
-    const drawFrame = () => {
-      ctx.clearRect(0, 0, imageSize.width, imageSize.height);
-      ctx.drawImage(image, 0, 0, imageSize.width, imageSize.height);
-      snowflakes.forEach((flake) => {
-        flake.update();
-        flake.draw(ctx);
-      });
-    };
-
-    for (let i = 0; i < 20; i++) {
-      drawFrame();
-      gif.addFrame(canvasRef.current, { copy: true, delay: 200 });
-    }
-
-    gif.render();
-  };
-
-  const createSnowflakes = () => {
-    const flakes = [];
-    for (let i = 0; i < 600; i++) {
-      flakes.push(new Flake(snowflakeSize, snowflakeSpeed, snowflakeColor));
-    }
-    return flakes;
-  };
-
   return (
     <div>
       <input type="file" onChange={handleImageChange} />
@@ -226,8 +176,15 @@ const SnowEffectOnUploadedImage = () => {
             onChange={() => setUseFlakeImage(!useFlakeImage)}
           />
         </label>
-        <button onClick={generateGIF}>Generate GIF</button>
-        {/* GIF 생성 버튼 */}
+        <GenerateGIFButton
+          canvasRef={canvasRef}
+          imageSize={imageSize}
+          image={image}
+          Flake={Flake}
+          snowflakeSize={snowflakeSize}
+          snowflakeSpeed={snowflakeSpeed}
+          snowflakeColor={snowflakeColor}
+        />
       </div>
     </div>
   );
